test: consolidate snakefile util tests into test_snakefile_util.py
Merge the branch's unstable-gating/aggregator tests (formerly
tests/test_snakefiles_util.py) into main's tests/test_snakefile_util.py
so src/snakefiles/util.py has a single test module.
